+++ features/prices/components/Price/Price.tsx
@@ -1,7 +1,7 @@
 import { useLocale, usePrices, type PriceCoingecko } from '@/features/prices'
 import { PriceChange } from './PriceChange'
 import styles from './Price.module.css'
-import { formatFiat } from '@/lib/utils'
+import { formatFiat } from '@/lib'
 
 export function Price({ price }: { price: PriceCoingecko }) {
   const { isValidating, isLoading } = usePrices()
