+++ lib/fetch.ts
@@ -1,0 +1,5 @@
+export async function fetcher(url: string) {
+  const res = await fetch(url)
+  if (!res.ok) throw new Error('Failed to fetch')
+  return await res.json()
+}
